added finding maximum number by using float variables

+++ src/com/bridgelabz/generics/Maximum.java
@@ -2,9 +2,9 @@ package com.bridgelabz.generics;
 import java.util.*;
 
 public class Maximum {
-	public static int num1;
-	public static int num2;
-	public static int num3;
+	public static float num1;
+	public static float num2;
+	public static float num3;
 
 	public static void main(String[] args) {
 		
@@ -12,11 +12,11 @@ public class Maximum {
 		MaximumTest maximumInteger = new MaximumTest();
 		System.out.println("welcome to find maximum Number");
 		System.out.println("Enter first Number:");
-		num1 = sc.nextInt();
+		num1 = sc.nextFloat();
 		System.out.println("Enter second Number:");
-		num2 = sc.nextInt();
+		num2 = sc.nextFloat();
 		System.out.println("Enter third Number:");
-		num3 = sc.nextInt();
-		System.out.printf("maximum Number in %d,%d,%d is %d:", num1,num2,num3,maximumInteger.maximum(num1, num2, num3));	
+		num3 = sc.nextFloat();
+		System.out.printf("maximum Number in %.2f,%.2f,%.2f is %.2f:", num1,num2,num3,maximumInteger.maximum(num1, num2, num3));	
 	}
 }
